Fixed potential circular conflicts.
