+++ af-location-service/src/main/java/com/revature/service/BuildingService.java
@@ -20,8 +20,8 @@ public interface BuildingService{
 	BuildingDetailsDto getBuilding(int index) throws Exception;
 	void updateCity(int index, String city) throws Exception;
 	void updateNumberOfFloors(int index, int floors) throws Exception;
- 	void addRoom(int index, Room room) throws Exception;
- 	void deleteRoom(int indexBuilding, int indexRoom) throws Exception;
+// 	void addRoom(int index, Room room) throws Exception;
+// 	void deleteRoom(int indexBuilding, int indexRoom) throws Exception;
 	void updateBuilding(int index, BuildingRequestDto dto) throws Exception;
 
     List<BuildingDto> getBuildingsByLocation(int index);
