+++ af-location-service/src/main/java/com/revature/service/LocationServiceImpl.java
@@ -19,9 +19,7 @@ public class LocationServiceImpl implements LocationService {
 
 	@Autowired
 	private LocationRepository locationRepository;
-	
-  	@Autowired
-	private BuildingServiceImpl buildingService;
+
 
 
 	@Override
@@ -136,15 +134,6 @@ public class LocationServiceImpl implements LocationService {
 	public void deleteLocation(int index) {
 
 		locationRepository.deleteById(index);
-
-	}
-
-	@Override
-	public void addBuilding(int index, BuildingRequestDto buildingRequestDto) {
-
-		Location location = locationRepository.findById(index).get();
-		buildingService.createBuilding(buildingRequestDto, location);
-		locationRepository.save(location);
 
 	}
